Ajuste no ccontroller e limpeza de codigo
