--- Spring/src/main/java/com/prateleira_inteligente/controllers/PrateleiraController.java
@@ -1,6 +1,0 @@
-package com.prateleira_inteligente.controllers;
-
-public class PrateleiraController {
-// codigo no pc de casa, verificar
-// verificar primeira linha também
-}
