--- Spring/src/main/java/com/prateleira_inteligente/dto/PrateleiraDTO.java
@@ -1,5 +1,0 @@
-package com.prateleira_inteligente.dto;
-
-public class PrateleiraDTO {
-// codigo no pc de casa, verificar
-}
